+++ cli/commands/add.ts
@@ -17,7 +17,6 @@ import { checkRequirements } from "../check-requirements.js";
 import { syncLocalCache } from "./install/sync-local-cache.js";
 import { notifyInstalls } from "../notify-installs.js";
 import { resolvePackages } from "../resolve-packages.js";
-import { stripRangePrefix } from "../semver.js";
 
 type AddOptions = {
   verbose?: boolean;
@@ -109,11 +108,9 @@ export async function add(
       process.exit(1);
     }
   } else if (!pkgDetails.path) {
-    let res = await installMopsDep(
-      pkgDetails.name,
-      stripRangePrefix(pkgDetails.version),
-      { verbose },
-    );
+    let res = await installMopsDep(pkgDetails.name, pkgDetails.version, {
+      verbose: verbose,
+    });
     if (res === false) {
       return;
     }
@@ -148,5 +145,6 @@ export async function add(
       `${pkgDetails.name} = "${pkgDetails.repo || pkgDetails.path || pkgDetails.version}"`,
   );
 
+  // check conflicts
   await resolvePackages({ conflicts: "warning" });
 }